requirements.txt
# Framework Principal
streamlit==1.28.1

# Análise de Dados
pandas==2.1.3
numpy==1.25.2
plotly==5.17.0

# Modelos de Previsão
prophet==1.1.4
statsmodels==0.14.0

# APIs e Requisições HTTP
requests==2.31.0
urllib3==2.5.0

# Configuração e Logging
pyyaml==6.0.1
python-dateutil==2.8.2

# Banco de Dados
psycopg2-binary==2.9.7
sqlalchemy==2.0.23

# Utilitários
python-dotenv==1.0.0
schedule==1.2.0

# Desenvolvimento e Testes
pytest==7.4.3
pytest-cov==4.1.0

# Qualidade de Código
black==23.11.0
flake8==6.1.0

